Restructure all package and modules.
Now we have a project with modules: rule-engine, itr-rule-engine, json-rule-engine and plugin. This restructuration was necessary in order to improve the project's clarity.

Now we have a test directory for each module that allow us to test individual behaviours. It helps us to understand better how itrules works in different situations.